Add MAX distance for record capture
Add MAX distance for record capture
Bugfix - max altitude shows max altitude rather than current altitude :)
Clearly never used ...!
Add max data menubuffer with 32 bit to avoid some overflow scenarios
Gui install instaructions

+++ MW_OSD/Serial.ino
@@ -439,7 +439,7 @@ For sub-command 3 (draw string):
 
   if (cmdMSP==MSP_COMP_GPS)
   {
-    GPS_distanceToHome=read16();
+    GPS_distanceToHome=(uint32_t)read16();
 #ifdef I2CGPS_DISTANCE
     gpsdistancefix();
 #endif
